requirements.txt
PyMySQL==0.9.3
python-telegram-bot==11.1.0
SQLAlchemy==1.3.1
